Update patch set 1
Patch Set 1:

It does not download deps? Never seen this issue myself, strange. Sounds like the actual error is somewhere else and --force simply ignores it and moves on, no?

Anyway, if this helps, --force should be removed as suggested but being able to re-run the script can save a lot of time so I'd like to keep it possible. Would maybe a `puppet bla bla || puppet bla --force bla` work? i.e. run without --force and if that fails, try again?

The cleaner fix (and cleaner as my old fix, too) would be to check whether the module is installed. If it is installed already, the line can be omitted. If it isn't, the --force is not necessary.

Patch-set: 1