Update patch set 3
Patch Set 3: Code-Review+1

Jan, when this was started the final release was not done.
RC1 is actually the same code as the release.
I suggest to get this online and later upgrade to the stable release.

Patch-set: 3
Reviewer: JF Joly <1002709@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=+1